Add an isoline feature.
This adds a Math.log10 polyfill and removes the unused Math.sinh
polyfill.

Add simple isoline example.

Add an isoline tutorial.  This uses a data set of Oahu's elevation that
is lower resolution than the 'dense' set, primarily so that the tutorial
will run faster (and test faster).

This resolves #592.