Pass position to the value function.
The used function can be overridden for isoline and contour features.
This defaults to the value function returning a non-null, finite number.

Pass the calculated position to the value function.  This allows the
value to be modified based on the position.

Add the isoline label feature to the list of dependent features so that
something like `myfeature.visible(false).draw()` does what is expected.

When auto-spacing isolines, make sure that labels don't have an
excessive number of digits due to bad rounding.

Don't wrap across longitude if the data is more than 360 units wide,
since it probably isn't actually longitude in that case.