Ensure that all point position calls pass the data index.
We want position to be a function that takes `(data, dataIndex)`, but in
several places we were only calling it with `(data)`.  Also, data
entries don't have to be objects, so guard against that -- you can set a
point feature with data just equal to `{length: <number of points>}`,
and then generate the positions purely based on the index.