Modify layers for better nested support.
Layer's z-index should only be relative to their siblings, not all
layers.  Layers don't have to be children of the map; they can be
children of other layers.  Instead of enumerating the map layers for
determining siblings, enumerate the children of the layer's parent,
filtering to only those that are geo.layer instances.

This makes it easier to change the z index of a layer that has dependent
layers (such as the isoline feature's canvas layer for text).